BUG: Change static DICOM dict initialization to fix lookups.
This is a fairly large change, it removes the union from the
tag class (it always stores the group, element numbers and never
a single 32-bit key) changes the hash function.  Most importantly,
it makes it possible to simply look at vtkDICOMMetaDataDict.cxx
and see what tags are in what rows of the hash table.