Rework GetAttributeValue() and the vtkDICOMValue interface.
I want it to be convenient to get values from the meta data,
so a straightforward GetAttributeValue() method that returns a
vtkDICOMValue object makes sense.  It is also easier to wrap.

The method names like GetULongData() have been renamed to
more standard VTK-like names, e.g. GetUnsignedIntData().

The NumberOfValues member of vtkValue has been made to match
the documentation: it is no longer set to 1 for OW, OB, OF, etc.