COMP: Missing vtable because virtual method not defined.
Certain conditions must be met in order for the compiler/linker to
generate a vtable.  Apparently because I defined a virtual method
in DecoderBase but forgot to give its definition, gcc assumed that
the vtable would be generated in whichever translation unit actually
defined that method.  Marking the method as pure virtual fixed it.
However, I don't understand why this problem snuck up on me.  I've
been compiling the code for months, no changes to the compiler or
that particular source file, and then one day... vtable gone!