Fix a bug in the computation of the RAS position elements.
The position changes when the MemoryRowOrder of the DICOM reader
changes.  This should not be so, because the end result coordinate
system should be identical (i.e. the flip is just happening in the
reader instead of in vtkDICOMToRAS).  I fixed one bug, but there is
still another hiding somewhere.