Add a method to search for per-frame attributes.
Multi-frame DICOM is hard... attributes can be in the Per-Frame
function sequence, in the Shared functional sequence, or just in
the root like the "old faithful" DICOM files.  So it is nice to
have a method that searches in all these places, even if it isn't
very efficient (it exhaustively searches all the mini-sequences
within the Per-Frame and Shared sequences).

I've used this method to ensure that the spacing and orientation
are set correctly in the reader, even if there are multiple stacks
with different spacings and orientations.