Change the order for pixel spacing to make it more robust.
I didn't like the if/else approach of getting pixel spacing.  Just
because the file has SharedFunctionalGroupsSequence, that doesn't
guarantee that the PixelSpacing isn't stored in the root (well, I
guess it does according to the standard, but I'm paranoid about
files that might not follow the standard to the letter).