Add per-instances method to vtkDICOMDataElement interface.
Earlier I was trying to keep the vtkDICOMDataElement interface as similar
as possible to how it appears in the DICOM standard.  This was a bad idea,
the interface should reflect usage first and foremost.  So I removed the
GetVL() method because people should get VL from the Value object, and
added methods so that people can check if the value is per-instance and
then get the value for a particular instance.

Also, the iterator's "*" operator can return a reference for efficiency.