Allow null values in meta data to mark missing attributes.
With some series, not all instances in the series share the same
tags, i.e. some instance have attributes that others are missing.
By adding null values as placeholders for missing attributes, it
becomes possible to reflect the true structure of the whole series
with just a single vtkDICOMMetaData object.