Allow the generators to place restrictions on the pixel type.
The new method SetPixelRestrictions allows the subclasses to provide
restrictions on the pixel types that are allowed.  In addition to this,
there is now some code that computes the scalar range of the data and
sets the BitsStored to be the minimum allowed value that can contain
the data.

Another tag-along patch fixes removes the PixelAspectRatio if the
PixelSpacing is set.  This problem first appeared when I rearranged the
order in which the modules were generated.