Bump the version number to 0.3.5.
This release does not add any features, it solely fixes warnings
and build problems.  The most crucial fix is that it explicitly
links to the VTK 6 libraries that are required, instead of using
the VTK_LIBRARIES variable which may contain unwanted text.  When
VTK was built with Qt5, the Qt libraries were listed in the form
Qt5::Core which was causing link problems (vtk-dicom does not use
Qt)