Use fread instead of ifstream, for performance.
I used a profiler on my Mac, and found that vtkDICOMParser spent
four times as much time in ifstream::read() as it did in fread().
This is when reading files that are already cached in memory, and
if they had actually been hitting the disk the difference might have
been less.