Add support for the GB18030 Chinese character set.
This should support most, but not all, of this particular character
set.  The linear table that converts 4-byte GB18030 sequences into
unicode is incomplete, but it should suffice for display of text.
The conversion only produces values under U+FFFF for now, private
codes are used in preference to standard codes that are above U+FFFF.