Explicitly check ranges for multi-byte iso-2022 codes.
The main benefit of this is that it allows the space character
to be interpreted as a single-byte character in iso-2022-jp text.
Otherwise, the trailing space that is mandatory padding at the end
of odd-length strings in DICOM would be interpreted as one half of
a corrupted multi-byte character.