Fix the java dependencies.
When parallel builds (make -j 2) were used, the build would fail
with the complaint that vtkDICOMHierarchy.txt was needed by one of
the .java build commands.  To fix, I made vtkDICOMHierarchy.txt one
of the outputs in vtkWrapHierarchy.cmake.  This meant copying the
whole vtkWrapHierarchy.cmake file to change just a single line.