Add another 8-bit field in vtkDICOMValue for overflow.
This is specifically for values that use delimiters instead of fixed
lengths, such as OB or UN VRs that store encapsulated data.  By adding
an 8-bit overflow byte for the 32-bit NumberOfValues field, it becomes
possible to store encapsulated data with sizes up to 2^40 bytes (1TB).
Values that use fixed lengths are still limited to 2^32-2 bytes (4GB)
as per the standard.