Give base attributes higher priority than those in private sequences.
This is necessary because of the way that Philips builds their enhanced
multi-frame files.  They have private sequences within the PerFrame and
Shared sequences that hold all of the legacy data.  Obviously, we don't
want this legacy data to be returned if the base data set has those same
attributes.  For example, the Philips sequence stores a legacy InstanceUID,
and we never want to return it in preference to the real InstanceUID.