Add sophistication to slice spacing computation.
Compute the slice spacing to be the average slice spacing after culling
outliers.  This makes the spacing computation robust when a series is
missing slices.  Unfortunately, everything past a missing slice is simply
shifted by one slice, which is not the right thing to do... in a later
commit, the gaps should be filled with black slices or perhaps with
interpolated slices.