Qualify when meta header used for query key matching.
Sometimes I want to query for meta header elements when reading a
file (e.g. when scanning for compressed files), but in general if
the query data set has a full meta header (e.g. because it was itself
read from a file), then we don't want to use its contents as query
keys.