Add new vr "UR" for URIs and URLs.
This new vr replaces some instances of UT in the dictionary.  Unlike UT,
it is always restricted to the Default Character Set, and it cannot contain
a backslash.  However, it is not allowed to be multi-valued, probably for the
sake of compatibility with old systems that will interpret it as UT.