Put source code in the Source subdirectory.
This is a bit of housekeeping that I've wanted to for a long time.
Originally, when the vtk-dicom project started, I'd intended it to
exist as a directory within a large project.  Now that it has grown
to be a substantial project in its own right, it requires a nested
directory structure.