Add execdir primary to dicomfind.
This primary is very useful when the directory names have spaces,
because it ensures that the executed command will not see those
spaces.  Also, if a series has been duplicated (the same series
appears in multiple directories), then each duplicate will be
independently executed upon.  Conversely, a series that is split
across directories will be executed in parts.