Unrecognized explicit VRs should have 4-byte VL.
This rule has been part of the standard since at least 2004.
The modified code in the DICOM parser will check the dictionary if
an unrecognized VR is encountered, and if the tag is not in the
dictionary, the VR will be treated as "UN".

I include a secondary rule for compatibility with badly-converted
data.  If the VR bytes do not consist of two capital letters, then
the data element is assumed to be implicitly encoded.