Add a filter to apply the palette to an image.
This is the first of what will potentially be a new breed of filters
that utilizes the DICOM meta data.  Note that the current implementation
is incomplete, it simply gets the lookup table from the first file in
the series that was used to construct the volume, instead of checking
whether each file had a different lookup table.