Fix VR problems with empty value constructor.
It was missing OB, and was mishandling OX and XS by actually using
OX and XS as the VR (even though they aren't legal, except in the
dictionary).  Now OX becomes OW and XS becomes SS, which is the
same as the other constructors.  The only time OX becomes OB is when
the value is explicitly constructed from an "unsigned char *", and
the only time XS becomes US is when the value is explicitly constructed
from "unsigned short *".