Change XS, OX, SpecificCharacterSet resolution in parser.
In implicit data sets, a dictionary VR of XS is resolved to SS or
US depending on what PixelRepresentation (signedness) is set to.
Previously, if parsing an item, PixelRepresentation was only searched
for within the item.  Now, it is searched for within the item and
then within the base data set.  This same approach is also used to
resolve OX to OW or OB.  This should improve accuracy when parsing
implicit data.

A similar scoping is used to determine the SpecificCharacterSet
within an item.