Remove asserts for invalid tag paths.
What should happen if someone tries to use a tag path where the head
of the path is not a sequence?  Previously, asserts were used in some
places and not in others.  I've removed the asserts.  Now, if someone
uses an invalid tag path, nothing happens (an attept to set an attribute
with an invalid path does nothing, an attempt to get an attribute with
an invalid path returns an invalid value).