The dicompull output tool should assume UTF8 encoding.
I've looked at the KWSys MakeDirectory function that dicompull calls,
and it appears to assume "local encoding".  On linux and OS X, this is
usually (fingers crossed) utf-8 for modern installations.  On Windows,
ideally the utf-8 should be converted to utf-16, but I don't think that
KWSys is doing that, it does not seem to assume that the local 8-bit
encoding is utf-8.