The vtkDICOMDirectory search skipped some subdirectories.
If the directory tree was not searched in lexical order, then some
subdirectories could be skipped.  This was not seen on OS X, which
iterates through directory listings in lexical order, but was seen
on other platforms.  Even on OS X, it could be seen in cases when
the hfs case-insensitive sorting did not match the case-sensitive
sorting that would have caused the bug to be hidden.