Use loose matching with an Osirix database file.
The Osirix .sql file contains "cleaned" values of the DICOM attributes,
for example commas are replaced with spaces and many other such
substitutions are performed.  The way that strings are cleaned has
changed between different versions of Osirix.  To accomodate this,
a loose comparison has been added to vtkDICOMDirectory when using
the .sql file.