Fix documentation for GetCharacterSetString.
The purpose of this fix is to clearly state the conditions under which
this method will return an incomplete set of code values.  To get a
complete set of values, it will be necessary to add a similar method
that takes an encoded string as a parameter and identifies the ISO 2022
escape codes that are present in the string.