Allow ISO 2022 mixture of kanji with half-width katakana.
The escape code <ESC>)I places ISO IR 13 into G1 (high bit set) so if
either of the kanji charsets are currently active in G0, they should
continue to be active.  I'm doubtful that <ESC>)I occurs "in the wild"
in DICOM or, in fact, in any text anywhere so this is a pedantic fix.