Add command-line option -u for specifying UIDs in a file.
It's not uncommon to use dicomtocvs to create a spreadsheet for all the
DICOM series, and then to want to use the UIDs from that spreadsheet
with dicompull.  This change allows the user to simply paste all of the
UIDs into a file, and ask dicompull to grab those series.