Use CP65001 on Win32 for command-line tools.
This is the UTF-8 code page for the Windows console.  All programs that
define their main() function with mainmacro.h use UTF-8 internally, and
all output is printed as UTF-8.  The console should therefore be set to
accept their output.  Currently, on UNIX, we assume that a UTF-8 locale
is being used.