Allow NumberOfFrames query with dicomtocsv.
The dicomtocsv program uses vtkDICOMMetaDataAdapter to search the
SharedFunctionalGroupSequence and PerFrameFunctionalGroupSequence
of enhanced mult-frame files.  This adapter hides the NumberOfFrames
attribute of the files and instead reports the NumberOfFrames as
the number of instances, in order to make the file "look" like an
entire series.  We do not want this behavior in dicomtocsv, so we
must sidestep the adapter when querying NumberOfFrames.