Allow the adapter to choose from a series of data sets.
This change adds an index to the constructor for vtkDICOMMetaDataAdapter.
In this way, if the adapter is constructed from the meta data for a
series of enhanced multi-frame data sets, the user can choose which one
of those data sets to view.  Previously, only the first of such
multi-frame data sets could be viewed through the adapter.