Add vtkIOMySQL to the link libraries if present.
This library provides factory classes (i.e. plug-in implementations)
for the classes in vtkIOSQL, which in turn are used by the class
vtkDICOMDirectory for reading OsiriX sql database files.

For most systems, vtkIOMySQL is linked transitively as a dependency,
but the the VTK .deb package avoids transitive linking, so on Debian
vtk-dicom must explicitly link all factory classes when building
against an apt-get installed VTK.