Make dicomfind {} substitution more flexible
When -exec was used with dicomfind, every occurrence of {} is replaced
by one of the found files.  Previously, this replacement only occurred
if an argument was exactly "{}".  Now, a search-and-replace is performed
such that every occurrence of {} within the argument is replaced by the
filename.  This is done for compatibility with the BSD "find" command.