Perform hangul composition from 8-byte jamo sequences
KS X 1001 encodes only 2350 of the modern Korean syllables, and allows the
remaining syllables to be coded as 8-byte sequences.  This patch decodes
those sequences to generate the proper unicode characters.

If the sequence encodes one of the 2350 syllables that exists in KS X 1001,
then the sequence is written out in full via the Hangul Jamo Compatibility
codes rather than being converted to a character.  This is necessary to
allow the sequence to round-trip back to KS X 1001.

Note that if the leading consonant or the vowel component of the syllable
is filler (i.e. empty), then unicode does not provide a pre-composed form,
and the character is written to unicode in decomposed form with two or
three combining codes.