Ensure bad chars and nulls are properly handled
If the trailing byte of a multi-byte sequence is invalid, then the
decoder should never "eat" the trailing byte.  It should always leave
the trailing byte on the stream, because the trailing byte could be an
important character.  In particular, we certainly never want to eat a
NULL and emit something other than NULL.