Convert unknown charsets as ASCII when possible
For unrecognized character sets, we can generally assume (with exceptions
noted below) that (0x00,0x7F) encodes mostly the same characters as ASCII.
Also, we should always assume that the control codes (0x00,0x20) and (0x7F)
are the same as US ASCII.

The exceptions are multi-byte encodings that use codes within (0x21,0x7E)
as either the leading or trailing character.  This includes, in particular,
some ISO-2022 multi-byte code, especially the iso-2022-jp family.  So I
have added a check for all such ISO-2022 codes to avoid converting that
range to ASCII when they are active.

There are some non-ISO-2022 multibyte encodings where the trailing byte
can be in the range (0x21,0x7F), so this patch will potentially allow
such trailing bytes to be erroneously interpreted as ASCII if such codes
occur and are not recognized.  Fortunately, we already recognize many of
these codes: GBK, GB18030, CP949 (as an extension to EUCKR), and Shift-JIS
(as an extension to JIS X 201).