Avoid eating backslash within escape codes.
Since the backslash is a delimiter, it must not be consumed when an
escape sequence is being parsed.  Also, if an escape code cannot be
parsed, then it should not be consumed.

Any valid escape sequences that begin with ESC, but which are not
ISO 2022 G0 or G1 designating sequences, will be passed verbatim
through the character set conversion.  The CSI code (0x9B) shall
not be recognized as an escape sequence initiator, because it can
appear as part of a character within some multi-byte character sets.