When rescaling, prefer only changing signedness.
The previous code was changing the output type to "float" for many
CTs simply because a shift of -1000 was changing the data range
(this occurs if BitsStored = BitsAllocated).  The new code will
only change the signedness (e.g. "short" to "unsigned short"), even
though this might lead to underflow or overflow, unless the rescale
results in fractional values.  Any underflow or overflow will be
clamped.