Improve matching of written slices to input slices.
Previously, the data-set generator would only capture slice-by-slice
meta data from its input if all of the input slices shared the same
orientation as the data to be written.  This was too strict, since
the input series could consist of multiple stacks, or a stack plus a
locator image.  Slice-by-slice matching should be possible as long as
each output slice matches an input slice; it is no concern if there
are leftover input slices after the matching operation.