Add --min-value and --max-value to dicomtocsv.
These options allow you to retrieve the maximum or minimum value
of an attribute across the series.  The purpose is to find the
maximum InstanceNumber so it can be compared to the number of
files in the series (if they are not equal, then it's likely that
a file is missing).