Make all query data available in vtkDICOMDirectory.
This change makes it so that all the attributes that are listed in
the query are available through the GetMetaDataForSeries() method.
It comes with a cost, however, because it means that if a DICOMDIR
or sql file is used to index the files, if the index doesn't have
enough information to satisfy the query, then all of the files in
that series are scanned rather than just the first file in the
series as was done previously.