Fix local timezone bug in DT generation.
When converting back and forth between UNIX time and DT strings,
if the timezone was not explicitly given, then the current local
timezone offset was used.  This was incorrect, since the "current"
offset does not always match the "historical" offset due to
daylight savings time and other historical offset changes.