Use system sqlite if building remote module.
This is messy, but the use of sqlite3 is optionally, I am not sure
how to optionally put it in module.cmake for a remote build.  So,
for a remote build, I'll stick with the system sqlite3 since the
module system will ensure that the vtksqlite3 headers aren't in the
path if vtksqlite3 is absent from module.cmake.