Improve dicomtonifti error reporting if no input.
After these changes, if dicomtonifti is not in batch mode but is
given no input files, it will print an error message.  Also, if
not in batch mode and given files directly on the command line
(not in a subdirectory) it will attempt to read them even if they
cannot be identified as DICOM.

Also, this change will cause the files given on the command line
to be sent to the reader with no sorting: it will be completely
up to the reader to sort them.