Copy attributes that aren't in SOP base class
Most DICOM files that exist "in the wild" are instances of
extended SOP classes, which means that they use one of the
SOP Class UIDs defined in the standard, but have additional
attributes as declared in their conformance statements (assuming
that the conformance statement is complete and up to date).
Unfortunately, I don't have copies of all the conformance statements,
or a way of matching every DICOM data set to a conformance
statement.  But I can add an option to copy all of the extended
attributes that are present.