Don't short-circuit parser on query failure
The short-circuit behavior was implemented so that dicomfind would
be faster when looking for matching files.  Unfortunately, not all
users expected this behavior.  In particular, dicomtocsv expected
all of the queried meta-data to be returned, regardless of whether
the query succeeded or failed.  The short-circuiting caused a major
performance loss because the SeriesInstanceUID was queried but was
not being returned because it came after a short-circuit failure.

A query serves two purposes: 1) it specifies the attributes to be
returned and 2) it gives match patterns for some attributes.  It is
wrong for (1) to fail because (2) doesn't match for some attribs.