Speed up search for duplicate UIDs
The search for duplicate IDs is an inner loop, so its efficiency
matters.  The previous code kept the vector sorted through the
use of insert(), which requires moving elements around.  The new
code keeps a separate vector of keys and indices, which is much
cheaper to keep sorted and to search.
This improves efficiency, because sorting a vector of pointers is
faster than sorting the original vector of objects.