Fix error when Database.sql-wal is missing for OsiriX
Sometimes the write-ahead-log for OsiriX is deleted accidentally or
intentionally (e.g. when the log is empty).  Unfortunately, opening
the Database.sql failed unless Database.sql-wal either existed or
could be created, and creating the file is impossible if the database
is read-only.

The solution is to use the "immutable" flag when opening the database.
This can be easily done opening the database via a URI and specifying
the flag in the URI.