Fix edge case for invalid VR on multiplex values
When some files in a series contained an attribute that was
missing from the first file in the series, the VR for the
multiplexed data element was not set, since the VR comes from
the first file in the series.  After this commit, the VR will
come from the first file that has the attribute.