Fix overflow when creating OW value from string
For OW, OL, OV, OF, and OD the value constructor allows pointers
of any numerical type and simply does a memcpy.  If the type of
the pointer was less than the value type size, then the allocated
space could be less than the memcpy, causing a buffer overlow.
This fix ensures that the memcpy size is the size of the buffer.