Fix infinite loop caused by garbage at end of file.
If between 1 and 7 bytes of "padding" (zeros or otherwise) were
present after the last data element in the file, the parser would
go into an infinite loop because it could not see the end-of-file
nor could it read the next element header.

The fix is to advance the position to the end of the file when
CheckBuffer fails.  This means any tail padding that is too small
to be a data element is silently ignored.