Fix the tilted-gantry PatientMatrix for BottomUp
The default "BottomUp" setting of the reader's "MemoryRowOrder"
parameter caused an incorrect PatientMatrix to be computed for
tilted-gantry images.  The third column of the matrix was being
set to the slice normal, instead of the vector along which the
gantry moved between slices.  Also, there was an error where not
all of the slice positions were being used to compute the vector.