Fix three GB2312/GBK compatibility mappings
These are for characters that are encoded twice by GB18030: first to
PUA code points in the area that GB18030 shares with GBK and GB2312,
and to non-PUA code points within the region of GB18030 that is a
linear mapping to Unicode.  Three of ten vertical punctuation
compatibility mappins were incorrect.