WIP: add --directory-only to dicomfind, dicompull
This option allows these tools to explicitly limit the search to
whatever metadata is present in the DICOMDIR (or sqlite) file.

This is a WIP because I'm not sure that this option is a good idea.
For dicomtocsv, the --directory-only option is used to ensure that
the output metadata comes only from the DICOMDIR, which is different
from requiring that the search only uses the DICOMDIR.

Ideally, the default behavior (only go to files for metadata not
present in the DICOMDIR) would be adequate.  I'll have to check
vtkDICOMDirectory to see if the files are being scanned unecessarily.