Add --directory-only to dicomfind, dicompull
This option allows these tools to explicitly limit the search to
whatever metadata is present in the DICOMDIR (or sqlite) file.
Ideally, the default behavior (only go to files for metadata not
present in the DICOMDIR) would be adequate, but vtkDICOMDirectory
seems to have a bug that causes it to scan the files unnecessarily,
and I haven't yet tracked down that bug.