Fix module import in TestDICOMPython.py
The CMakeLists.txt set PYTHONPATH to the LIBRARY_OUTPUT_DIRECTORY
of the module, which was incorrect.  For proper testing, the
PYTHONPATH needs to be set to the directory in which the package
resides (e.g. site-packages), not to the directory in which the
module itself resides (which is within the package itself).