Add latin9 names ISO_IR 203, ISO 2022 IR 203
In DICOM 2021d, latin9 became one of the officially recognized
character sets for DICOM, since it is sometimes used instead of
latin1 within the EU.  Latin9 adds several useful characters that
were absent from latin1 but present in Windows-1252.  Most notable
are the Euro currency symbol, and the oe ligature that is used in
many French words.