Make dicomtocsv default to consensus, not first
Previously, for each queried tag, dicomtocvs would by default
report the value for that tag for the first image in the series
where the value wasn't missing or empty.  However, when the first
in the series was a localizer, many image-related values differed
between the localizer and the remainder of the series.

Now the default is to look for runs where the queried tag has the
same value for multiple images in the series.  The value that has
the longest run is the one that is reported (with earlier taking
priority over later runs, if multiple runs have the same length).