WIP: add code to dicomtocsv to ignore localizer
If the first or final image in a series has a different orientation
than the second image in the series, then that image will be assumed
to be a localizer and will be skipped by dicomtocsv.  Not that if
both the first and final image have different orientations than
the second image, neither will be skipped.  Ditto if orientation
is absent from the images.