Don't build examples when building remote module
When vtkDICOM is built with VTK as a remote module, the topological
dependency sort requires that the module file lists all of the
dependencies.  However, the modules have their own specific
dependencies that we don't want in the library's module file.
So we disable the examples when building as a remote module.