For ISO 2022 IR 13 do backslash,tilde->yen,macron
This was already done for JIS X 0201 on its own, but not when
JIS X 0208 and JIS X 0212 were available via ISO 2022.  Instead,
for ISO 2022, backslash and tilde were stored in JIS X 0208 and
JIS X 0212 respectively.  This was antiquated behavior that does
not match the modern intepretations of these character sets, where
JISX 0208 and 0212 use the wide versions of those characters.