Fix implemenation of CP 246 (explicit UN)
This CP for Part 5 Section 6.2.2 added the important line of text,
regarding conversion of Implicit VR to Explicit VR for attributes
not in the dictionary (e.g. private attributes):

  "In the case of Undefined Length Sequences, the Value shall remain
  in Implicit VR form."

The parser was using Implicit VR to read these sequences even when
the length wasn't undefined, causing failure for some files.