Handle duplicates in tpl private dict parser
The tpl files from dicom3tools contain duplicate entries for some
private dictionaries.  Often one of the duplicates has a Keyword
while the other does not, in which case we prefer the one with
the Keyword.  This is done on the assumption that the one with
the Keyword was taken from a conformance statement, while the one
with no Keyword was not.

In three instances, the duplicates were caused because the same
element number was used twice in a row, and not incremented:

"BRIT Systems, Inc." (0021,0034) -> (0021,0035)
"SIEMENS MR VA0  GEN" (0021,0072) -> (0021,0073)
"PRIVATE_CODE_STRING_1001" (1001,00c3) -> (1001,00c4)

In another instance, only the VM was different, and since the
Keyword was plural, the one with VM=1-n rather than 1 was chosen.

"Sound Technologies" (f001,000f) VM=1 -> VM=1-n