Try to keep block when setting private elements
When a private element (gggg,xxee) is resolved to (gggg,yyee)
prior to writing, the code now tries to use the private block
yy to xx if it is available, but if xx is not available, then
yy will be set to the first available block.

The main purpose for this is when private elements are set on
the command line via "[creator]0019,1201", whence block "12"
will be preferred over "10" or any other.  This places more
control in the hands of the user.