Fix queries where private element has no creator
This is specifically for when the query (which is, itself, a tiny
dataset) contains private elements for which the creator is not
known.  Since private tags have different meanings for different
vendors (and sometimes different meanings for the same vendor),
such queries must be used with caution.  Leaving out the creator
is really just a convenience for people familiar with datasets
produced by their vendor.