Don't require VR to match when doing a value match
This is needed in particular for queries on private datasets, where
the VR in the dataset might not match the VR in the query.  The
matching method that is used depends on the VR in the dataset
rather than the VR in the query, so wildcards and ranges will only
work if the dataset VR allows them to.

Since US to SS matches are now allowed, some code was added to make
sure that the sign bit of the compared values is not set.