Programs now do private query without creator or VR
Previously, "-k [creator]gggg,eeee" would be needed to query a
private element, or "-k [creator]gggg,eeee:VR" to additionally
supply the VR if the private tag wasn't found in a dictionary.
Now a creator-less private tag can be used, with or without a VR.
If the VR is not specified, a default VR of UT will be used, in
order to allow text-based matching of values.