Fix reading sql files on read-only filesystems
Even in read-only mode, by default sqlite expects write access to
the write-ahead-log for locking.  So if we don't have write access,
we boldly tell sqlite to assume the database is immutable and does
not require locking.  This allows us to read sql files that have
been archived on read-only filesystems.